requirements.txt
flask==3.0.2
python-dotenv==1.0.1
openai==1.12.0
opencv-python==4.9.0.80
numpy==1.26.4
torch==2.2.1
torchvision==0.17.1
transformers==4.37.2
pillow==10.2.0
gtts==2.5.1
cv2==4.9.0.80
timm==0.9.12